+++ src/main/kotlin/com/procurement/access/infrastructure/handler/v2/model/response/GetDataForContractResponse.kt
@@ -3,6 +3,7 @@ package com.procurement.access.infrastructure.handler.v2.model.response
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
+import java.math.BigDecimal
 
 data class GetDataForContractResponse(
     @param:JsonProperty("tender") @field:JsonProperty("tender") val tender: Tender
@@ -92,7 +93,7 @@ data class GetDataForContractResponse(
             @JsonInclude(JsonInclude.Include.NON_NULL)
             @param:JsonProperty("additionalClassifications") @field:JsonProperty("additionalClassifications") val additionalClassifications: List<AdditionalClassification>?,
 
-            @param:JsonProperty("quantity") @field:JsonProperty("quantity") val quantity: Int,
+            @param:JsonProperty("quantity") @field:JsonProperty("quantity") val quantity: BigDecimal,
             @param:JsonProperty("unit") @field:JsonProperty("unit") val unit: Unit,
             @param:JsonProperty("description") @field:JsonProperty("description") val description: String,
             @param:JsonProperty("relatedLot") @field:JsonProperty("relatedLot") val relatedLot: String
